added a check for delete

+++ TrainingProject/Application/Commands/StoreDepartments/DeleteStoreDepartmentCommand/DeleteStoreDepartmentCommandHandler.cs
@@ -4,12 +4,12 @@ using Microsoft.EntityFrameworkCore;
 using System.Threading;
 using System.Threading.Tasks;
 using TrainingProject.Core;
-using TrainingProject.Domain;
+using TrainingProject.Core.Exceptions.StoreDepartmentExceptions;
 using TrainingProject.tables;
 
 namespace TrainingProject.Application.Queries.StoreDepartments.DeleteStoreDepartment
 {
-    public class DeleteStoreDepartmentCommandHandler : IRequestHandler<DeleteStoreDepartmentCommandQuery, StoreDepartmentDomainModel>
+    public class DeleteStoreDepartmentCommandHandler : IRequestHandler<DeleteStoreDepartmentCommandQuery, StoreDepartment>
     {
         private readonly ApplicationContext _context;
         private readonly IMapper _mapper;
@@ -20,16 +20,19 @@ namespace TrainingProject.Application.Queries.StoreDepartments.DeleteStoreDepart
             _mapper = mapper;
         }
 
-        public async Task<StoreDepartmentDomainModel> Handle(DeleteStoreDepartmentCommandQuery request, CancellationToken cancellationToken)
+        public async Task<StoreDepartment> Handle(DeleteStoreDepartmentCommandQuery request, CancellationToken cancellationToken)
         {
             var storeDepartment = await _context.storeDepartments
+                .Include(sd => sd.Stands)
                 .FirstOrDefaultAsync(sd => sd.StoreId == request.StoreId && sd.DepartmentId == request.DepartmentId, cancellationToken);
             if (storeDepartment == null)
                 throw new StoreDepartmentNotFoundException();
+            if (storeDepartment.Stands.Count != 0)
+                throw new StoreDepartmentDeleteException();
 
             _context.storeDepartments.Remove(storeDepartment);
             await _context.SaveChangesAsync(cancellationToken);
-            return _mapper.Map<StoreDepartmentDomainModel>(storeDepartment);
+            return storeDepartment;
         }
     }
 }
